Change prefDef to use true spelling, not dict
This is probably better than using the dictionary spelling, because
different inflections are more likely or less likely to be particular
ambiguous words, and because using the dictionary form means that
deconjugations of the same string with different dictionary spellings
can have multiple preferred terms, which would be thrashed if the user
goes through the checkboxes after right clicking it. Of course, this
makes it so that users have to specify more preferred definitions
sometimes to make the right stuff show up, but that's how "known word"
works already, too. This is, of course, an experimental change.