Overhaul word splitter to reduce complication
The algorithm's concepts and a summary of it are explained in comments
before it. It has been restructured so that each intent in its design
has its own obvious single section of code. This is an invasive overhaul
and might have bugs that aren't in the tests because they're so
unexpected.